+++ R/aftreg.R
@@ -5,7 +5,7 @@
 #' @description Function to fit accelerated failure time (AFT)  models.
 #' @param formula an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.
 #' @param data data an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which function is called.
-#' @param baseline the chosen baseline distribution; options currently available are: exponential, weibull, lognormal and loglogistic distributions.
+#' @param baseline the chosen baseline distribution; options currently available are: exponential, weibull, lognormal, loglogistic and Birnbaum-Saunders (fatigue) distributions.
 #' @param dist alternative way to specify the baseline distribution (for compability with the \code{\link[survival]{survreg}} function); default is NULL.
 #' @param init initial values specification (default value is 0); see the detailed documentation for \code{init} in \code{\link[rstan]{optimizing}}.
 #' @param ... further arguments passed to other methods.
@@ -13,21 +13,27 @@
 #' @examples
 #' \donttest{
 #' library(survstan)
-#' fit <- aftreg(Surv(futime, fustat) ~ ecog.ps + rx, data = ovarian, baseline = "weibull", init = 0)
+#' fit <- aftreg(Surv(futime, fustat) ~ ecog.ps + rx, data = ovarian, baseline = "weibull")
 #' summary(fit)
 #' }
 #'
-aftreg <- function(formula, data, baseline = c("exponential", "weibull", "lognormal", "loglogistic"), dist = NULL, init = 0, ...){
+aftreg <- function(formula, data, baseline = "weibull", dist = NULL, init = 0, ...){
   if(!is.null(dist)){
     baseline <- dist
   }
   baseline <- tolower(baseline)
-  baseline <- match.arg(baseline)
-  mf <- stats::model.frame(formula, data)
+  baseline <- match.arg(baseline, survstan_distributions)
+  Call <- match.call()
+  mf <- match.call(expand.dots = FALSE)
+  m <- match(c("formula", "data"), names(mf), 0L)
+  mf <- mf[c(1L, m)]
+  mf[[1L]] <- quote(stats::model.frame)
+  mf <- eval(mf, parent.frame())
+  mt <- attr(mf, "terms")
   resp <- stats::model.response(mf)
+  X <- stats::model.matrix(mt, mf)
   time <- resp[,1]
   event <- resp[, 2]
-  X <- stats::model.matrix(formula, data = mf, rhs = 1)
   labels <- colnames(X)[-1]
   X <- matrix(X[,-1, drop = FALSE], ncol=length(labels))
   n <- length(time)
@@ -35,16 +41,17 @@ aftreg <- function(formula, data, baseline = c("exponential", "weibull", "lognor
   tau <- max(time)
   y <- time/tau
 
-  output <- list(call = match.call(), formula = formula,
-                 terms = stats::terms.formula(formula), mf = mf,
-                 baseline=baseline, survreg = "aft",
-                 n=n, p=p, tau=tau, labels=labels)
+
+  output <- list(call = Call, formula = stats::formula(mt),
+                 terms = mt, mf = mf, baseline = baseline, survreg = "aft",
+                 n = n, p = p, tau = tau, labels = labels)
 
   baseline <- switch(baseline,
-    "exponential" = 1,
-    "weibull" = 2,
-    "lognormal" = 3,
-    "loglogistic" = 4
+                     "exponential" = 1,
+                     "weibull" = 2,
+                     "lognormal" = 3,
+                     "loglogistic" = 4,
+                     "fatigue" = 5
   )
 
   stan_data <- list(time=y, event=event, X=X, n=n, p=p,
@@ -66,13 +73,15 @@ aftreg <- function(formula, data, baseline = c("exponential", "weibull", "lognor
   }
 
   output$residuals <- switch(output$baseline,
-                      exponential = -stats::pexp(nu, rate = pars[p+1], lower.tail = FALSE, log.p = TRUE),
-                      weibull = -stats::pweibull(nu, shape = pars[p+1], scale = pars[p+2], lower.tail = FALSE, log.p = TRUE),
-                      lognormal = -stats::plnorm(nu, meanlog = pars[p+1], sdlog = pars[p+2], lower.tail = FALSE, log.p = TRUE),
-                      loglogistic = -actuar::pllogis(nu, shape = pars[p+1], scale = pars[p+2], lower.tail = FALSE, log.p = TRUE)
+                             exponential = -stats::pexp(nu, rate = pars[p+1], lower.tail = FALSE, log.p = TRUE),
+                             weibull = -stats::pweibull(nu, shape = pars[p+1], scale = pars[p+2], lower.tail = FALSE, log.p = TRUE),
+                             lognormal = -stats::plnorm(nu, meanlog = pars[p+1], sdlog = pars[p+2], lower.tail = FALSE, log.p = TRUE),
+                             loglogistic = -actuar::pllogis(nu, shape = pars[p+1], scale = pars[p+2], lower.tail = FALSE, log.p = TRUE),
+                             fatigue = -extraDistr::pfatigue(nu, alpha = pars[p+1], beta = pars[p+2], mu = 0, lower.tail = FALSE, log.p = TRUE)
   )
   output$event <- event
 
   class(output) <- c("aftreg", "survstan")
   return(output)
 }
+